+++ SnackerEngine/src/Graphics/Meshes/Circle.cpp
@@ -12,12 +12,12 @@ namespace SnackerEngine
 	{
 		if (subdivisions < 3) subdivisions = 3;
 		std::vector<float> vertices;
-		vertices.reserve((subdivisions + 1) * 2);
+		vertices.reserve((static_cast<std::size_t>(subdivisions) + 1) * 2);
 		std::vector<unsigned int> indices;
 		indices.reserve(subdivisions * 3);
 		vertices.push_back(0.5f); vertices.push_back(0.5f);
 		for (unsigned int i = 0; i < subdivisions; ++i) {
-			float rad = static_cast<float>(i) / static_cast<float>(subdivisions) * 2.0f * M_PI;
+			float rad = static_cast<float>(i) / static_cast<float>(subdivisions) * 2.0f * static_cast<float>(M_PI);
 			vertices.push_back(std::cos(rad)*0.5f + 0.5f);
 			vertices.push_back(std::sin(rad)*0.5f + 0.5f);
 			indices.push_back(0);
